kcmodule: Remove SuspendHybrid from BatteryCriticalAction options
We now have SleepMode::HybridSuspend, to be used with
PowerButtonAction::SuspendToRam. Let's not lock ourselves into
also providing a PowerButtonAction that achieves the same thing
but needs a separate code path.

The global config KCM did not support selecting "Hybrid Suspend"
as an option in Plasma 5. Only since two months ago, i.e. in
pre-alpha versions of Plasma 6, did I expose this to the UI
via commit dfab5705. In hindsight, this was not a good idea
because it makes removing PowerButtonAction::SuspendHybrid harder.

Another follow-up commit will deal with removing this enum value
altogether. But for now, let's stop the bleeding by not releasing
a version of Plasma (even alpha) where the UI allows selecting it
in the global config KCM. (Note that for the profiles KCM, usage
of PowerButtonAction::SuspendHybrid has already been migrated
to PowerButtonAction::SuspendToRam + SleepMode::HybridSuspend.)

We'll still want this option in the long run, let's spend some time
later on to figure out how to properly expose it to the user.