suspendsession: reenter suspend after waking up due to AC state change or mouse move with lid closed
On some devices, plugging or unplugging AC power or moving the mouse while suspended will wake the system up, and it will not go back to sleep.

Before suspending, remember the state of AC, the suspension mode (to ram/to disk) and the reason for suspending. On about to resume, if AC state has changed or the reason for suspend was lid close and the lid is still closed, and the suspend method was not suspend-then-hibernate, then abort the resume and instead send the device back to sleep, in the mode appropriate for the new AC state.

BUG: 476492
CCBUG: 466025