Manually listen for value changes in TimeDelaySpinBox UI components
Prior to this change, the "Reset" and "Defaults" buttons in the
recently QML-ported profiles KCM would fail to reset the value
in time delay spinbox components when the duration is modified
through text edits. For example, using the keyboard to change
"after 10 min" to "after 15 min" followed by Enter would work,
but restoring it via "Reset" or "Defaults" would not.

Value changes through spinner arrow clicks didn't cause trouble
because the value binding to the settings object remains intact.
Somehow text edits are different though, and the binding breaks
when assigning a new value. Hence the manual new listener.