kcmodule: Fix flicker when switching between profile tabs
BUG: 478271

This commit is not tackling whatever root cause might be causing
the flickering (likely related to delayed layouting in FormLayout).

Instead, we resolve the flicker by reusing the same single form
for all tabs. So we don't switch between different sets of controls,
but we merely reassign values for the one remaining set of controls.

On the implementation side, the main thing that's required is a new
Connections object with onProfileSettingsChanged() handler, to call
the same reset logic that some components already use for handling
change signals of individual properties within profileSettings.
Controls that never break property bindings do not need any changes.