Remove the "Lock screen" option from "After a period of inactivity"
BUG: 478531

This is already configurable in the Screen Locking KCM and
having a separate second setting for the same thing is confusing.

One deleted line in the KCM, the rest is all migration code.
The migration takes care of:

* Any users that haven't yet migrated from Plasma 5.
* Plasma 6 Alpha & Beta users who run RC1 or RC2 at least once.

Users who jump from Plasma 6 Alpha or Beta straight to 6.0 or above
will have to manually change `AutoSuspendAction=32` in their
powerdevilrc to `AutoSuspendAction=0`.