daemon: On Linux, set backlight brightness with login1 instead of KAuth
This slightly increases the "frame rate" of brightness changes
on internal displays on Wayland, and reduces the need for
privileged helpers (in favor of widely available login daemons).

BacklightBrightness now uses BacklightSysfsDevice for reading
brightness (like BacklightHelper), but for writing we rely on
the org.freedesktop.login1.Session D-Bus interface.

The Linux KAuth backlighthelper is removed. We're going to assume
that Linux users at this point are either running systemd-logind
or elogind, both of which provide the D-Bus interface.

Note that the FreeBSD backlighthelper is still around and does not
depend on anything systemd-related.

CCBUG: 495545