kbd backlight: Fix double brightness restore on LidOpen-resume
This fixes the issue that on LidOpen-resume the brightness might get restored
two times. Once to the correct vallue and once incorrectly to 0.

To archive this it uses the same simple "don't restore to 0"-policy as is used
in dimdisplay.cpp, dpms.cpp, and handlebuttonevent.cpp.

Signed-off-by: Werner Sembach <wse@tuxedocomputers.com>

(cherry picked from commit aa2fa106)