ddcutil brightness setter with delay to reduce display EEPROM writes
EEPROM has memory writes limit, so requests to a memory must be reduced
as much as possible to not damage a monitor.

This commit introduce a timer which writes brightness to EEPROM only after
one second of the last call to set brightness. Interval of one second
is chosen because official the Dell app uses such value. It also removes the
brightness animations from ScreenBrightness since they are no longer needed
and make things more complicated.