ddcutil redetect support and dmps state monitoring (add support for ddcutil 2.1)
This is attempt to add redetect support and dmps state monitoring. To test this you need at least `ddcutil 2.1.0` because ~without it this project will not be compiled~ without it redetect and dpms will not work. If someone will be interested in this project I will finish this. If not I will close it. I will also look forward to your suggestions and feedback.

For now it is working ~but have some work to do~.

tested with 2 different dell displays and it works good.