daemon: Reduce unwanted brightnessChanged signals that we originated
In BacklightBrightness::onDeviceChanged(), the brightnessChanged
signal is emitted after filtering out udev events while the
brightness animation is running. Unfortunately, these signals
are received with a short delay compared to when the timer was
started, possibly because the KAuth helper runs on a different
timeline and definitely because udev brightness notifications
are not synchronous calls.

This commit introduces a more robust way of filtering out change
events that BacklightBrightness itself was asking for.

In essence, we will now filter out events where the brightness
moves toward the current animation target brightness
(dubbed "executed brightness" as it's not always animated).

There is a rub, in that the brightness helper can still continue
its animation e.g. toward higher values until it receives the
message to turn around toward the lower value that the user has
suddenly decided to request instead. In order to ignore these
overshooting outliers, we also introduce an "expected min/max"
pair of brightness values that gets initialized with the previous
brightness target.

In addition to reducing unwanted change signals, this new method
also allows BacklightBrightness to exactly track udev brightness
values and will correctly detect external changes even during
a brightness animation.