daemon: Calculate brightness logic with runtime minimum value
Instead of a hardcoded valueMin() for all displays via overloading,
valueMin becomes part of the BrightnessInfo struct and is set by
the controller classes directly.

This will allow us to pass a configured minimum value later.
For now, we're just moving the hardcoded value to the screen and
keyboard brightness controllers respectively.

Furthermore, the BrightnessLogic class now stores its state as a
single BrightnessInfo member variable, instead of mirroring each
of BrightnessInfo's members as a separate variable.