daemon: Introduce setBrightnessMultiplier() as better dimming API
Prior to this commit, DimDisplay used to track brightness values
of the only display it was told about by ScreenBrightnessController.
When dimming stopped, the original brightness value was restored.

This doesn't work very well when multiple displays are present,
each presumably with a different brightness value. DimDisplay will
have an even harder time with edge cases such as restoring
brightness values for a display that was temporarily unplugged.

ScreenBrightnessController is a better place for this complexity.
It can exempt displays from dimming, reset the newly introduced
brightness multiplier to 1.0 when a new display or external
brightness change is detected, and pretend that brightness is
still the same to avoid confusing other clients.

I'm a little uncomfortable with all the various flows within
onDetectorDisplaysChanged(), and ideally we should add autotests
for some of the expected scenarios, but at least it's all localized
in one function. The rest is pretty straightforward.

BUG: 476550