daemon: Introduce setBrightnessMultiplier() as better dimming API
Prior to this commit, DimDisplay used to track brightness values
of the only display it was told about by ScreenBrightnessController.
When dimming stopped, the original brightness value was restored.

This doesn't work very well when multiple displays are present,
each presumably with a different brightness value. DimDisplay will
have an even harder time with edge cases such as restoring
brightness values for a display that was temporarily unplugged.

ScreenBrightnessController is a better place for this complexity.
It can exempt unmanaged displays from dimming, reset the newly
introduced brightness multiplier to 1.0 when a different client
assigns fixed brightness values, and pretend that brightness is
still the same to avoid confusing other clients.