daemon: Only external brightness change signals from DisplayBrightness
DisplayBrightness implementations have so far needed to emit
a brightnessChanged signal for ScreenBrightnessController to
pick up on, both as a result of explicit setBrightness() calls
and if they pick up on external brightness changes.

This is awkward for DisplayBrightness to implement. They can emit
the brightnessChanged signal immediately when setBrightness()
is called, which is strictly speaking a lie, or they can emit it
only once brightness has actually been set, at the risk of causing
a UI component to reset its value when that component has already
moved on to a different value instead.

Neither is great, so let's scrap the brightnessChanged signal
that follows setBrightness(). ScreenBrightnessController can assume
that the brightness gets changed as requested, and will take over
the responsibility of emitting signals from within its own
setBrightness() function. If anything goes wrong, DisplayBrightness
implementations can notify the controller about the failure.

What remains is a a signal purely for changes that the controller
did not ask for. We call this externalBrightnessChangeObserved.

Nothing changes for users of ScreenBrightnessController, for the
time being. However, this change paves the way for a later
introduction of source client/context parameters without having
to burden DisplayBrightness implementations with forwarding those.
It might also make backlight brightness sliders a tad snappier.