kcmodule: Focus custom command field once it appears
Use a FocusScope as root element for the `RunScriptEdit` component,
and set initial focus within it to the `commandText` field.
Forcing focus after the corresponding checkbox is toggled on
will then work correctly.

Unfortunately, the containing Flickable does not scroll down to
a newly visible command field. That's a different issue though.