kcmodule: Switch to comboboxes for selecting idle timeout values
This follows the direction spearheaded by kcm_screenlocker,
in plasma/kscreenlocker commit 311db2a9. In addition to a few
select preset options, the user can also select a "Custom…"
option which presents a duration prompt dialog.

Using comboboxes has a few nice UX characteristics:

* It allows us to replace the checkbox of some timeout settings
  with a textual option called "Never", which is easier to grasp.
  Likewise, "Immediately" is more obvious than "0 sec".

* It increases the mouse area available for clicking, which makes
  the control easier to hit for users with low-precision pointer
  input devices.

* It sidesteps the problem of SpinBox ignoring the defined
  step size when scrolling with a high-resolution scroll wheel.

* By putting the SpinBox into a custom duration prompt dialog,
  it's more obvious that both seconds and minutes are supported
  for each setting. Some options also provide second-level presets.

The downside is that it now takes three extra clicks to enter
and confirm a custom timeout duration. Hopefully our users are
comfortable with that trade-off.

In order to implement this, the latest versions of the
ComboBoxWithCustomValue and DurationPromptDialog components
are copied from kcm_screenlocker. Several improvements to these
components will be fed back to kcm_screenlocker and hopefully
upstreamed to Kirigami (or Kirigami Addons) at a later time.

In addition, this commit introduces TimeDurationComboBox, a glue code
component combining two aforementioned components, to further
reduce the amount of boilerplate required in ProfileConfig.qml.

BUG: 482428
CCBUG: 482853