daemon: Track display ids for legacy brightness API calls
In ScreenBrightnessController, this just moves the logic for
determining the displays for legacy setBrightness() (i.e. the one
without displayId) from that method into onDetectorDisplaysChanged().

We also introduce a getter and a signal that should later make it
possible to move much of the legacy API functionality out into the
BrightnessControl action.

It's now easier to tell where exactly we rely on legacy semantics.