Added setting for battery conservation mode (Lenovo)
The setting is not consistently persistent, it resets occasionally
after power loss and/or reboot. It might be a good idea to
implement a toggle for restoring the setting on reboot.

The path is fixed for now, as there is no evidence of a device
that differs from it. At least I wasn't able to find one after
a quick search.

BUG: 441057
FIXED-IN: 6.1