notifications: show popup and sound for charge complete
Since the applet hides by default as soon the device is plugged in, it is not possible to see at a glance whether the battery is still charging or fully charged. To save the people who want to know this an extra click, make the existing Charge Complete notification visible and audible by default.

CCBUG: 486781