OSD: Close on focus lost
The OSD is closed when the active window is changed from the OSD to another window.

This happens if the user clicks outside the OSD.
This mimics the behaviour of modal dialogs.

This also happens with actions such as Alt+Tab.
It is currently quite annoying to close the OSD once the window has lost focus.
There is no close button and the window consists of ~90% click targets.