kcm: use window colorset for NavigationTabBar
The Header colorSet is used by default here, but that only looks good
when the page header is intended to visually merge with the window
titlebar. In this case the page header sits below a window header, and
using the Header colorSet for both results in a massive expanse of gray.
Use the Window colorset instead so that it uses more semtnaically
appropriate window colors.