applets/batterymonitor: Show power profile badges on the icon if possible
Currently we get complaints that the power profile icons sometimes
replace the battery icons, obscuring the battery level and charging
indicator.

This commit makes use of some new Breeze icons added in
frameworks/breeze-icons!376 to show a badge of the current power
profile in the corner of the battery icon when that profile is a
non-default one (i.e. don't show the user their default power profile
all the time; they already know what it is). This lets users see all
relevant pieces of information at once, which is quite nice.

When using other icon themes lacking these new icons, the widget will
fall back gracefully to only showing the battery icon and not showing
the power profile icons at all. This is not a catastrophe IMO since that
information is still available in the popup.

FEATURE: 483805
FIXED-IN: 6.2.0