kcmodule: update combobox capitalization
The KDE HIG were recently updated to clarify guidance on combobox
capitalization. Combobox text is now supposed to use sentence
capitalization. Most comboboxes already follow this scheme, but the ones
allowing the user to select the duration of inactivity do not.

This change updates the labels to conform to the new guidance.

HIG merge request at documentation/develop-kde-org!433