daemon: Introduce ScreenBrightnessController::adjustBrightnessRatio API
This new method is intended as alternative to adjustBrightnessSteps,
supporting arbitrarily small floating-point deltas. The part of the
delta which is not reflected in a discrete brightness value change
will be accumulated in a floating-point tracking error value and
added to the delta on the next adjustment call.

This corresponds to the current behavior of high-resolution
(touchpad) scrolling on the applet icon, but generalized to
multiple displays. For the time being, we only adjust the set of
legacy API displays. This can be expanded to the full set of displays
once the OSD popup API also supports multiple displays.

For consistency, this commit also reworks the implementation of
adjustBrightnessSteps to affect multiple displays at once. The first
display in the adjusted set that actually changes its brightness
will be used as reference for all other available displays, which
are adjusted by a similar delta.

This implementation makes it possible to adjust brightness in sync
between different devices, continuing to adjust even when a subset
of displays is already hitting the 100% or 0% boundary. Only when
all displays have hit their brightness boundary, no action will be
taken and no tracking error accumulated.