kcmodule: Change inactivity labels
The label for configuring the action to take when inactive (and the
duration required to count as such) currently reads as follows:
"After a period of inactivity: [Sleep] [After 15 minutes]"

This is somewhat confusing, as it reads like a sentence but does not
quite make sense as that. It implies that there are two durations,
one of them being of unspecified length, and the configurable duration
only begins to count after this. It could even be read in such a way
that if the user was inactive for a period of time, it will sleep
15 minutes later even if the user interacts with the computer again
during these 15 minutes.

This change modifies the wording to read:
"When inactive: [Sleep] [After 15 minutes]"
This rules out the incorrect reading, is much shorter, and makes the
text more similar to that of the surrounding options, which all use
a "When x, do y" form. Some related text is also updated to match this.

This requires changing the keyboard accelerator from Alt+F to Alt+I, as
there is no F in the new text.