kcmodule: Switch script text field and combobox
The entry field for scripts does not have an explicit label, leading to
potential confusion of the user if the text field is filled and the
placeholder text no longer visible.

This change swaps the order of the text field and the timer combobox,
and amends the text to indicate that after the specified time, the
contents of the field will be run.