kcmodule: Change 'execute' to 'run'
The HIG recommend avoiding the term "execute" as the word in some
circumstances has negative associations.

The powerdevil kcm already follows this in most cases, but there is
still a visible "execute" in the dialog for setting a custom duration
before scripts are run. This change updates this use to "run", and
adjusts some ancillary text like translation contexts for consistency.