kcmutil: Make script terminology more consistent
The labels, translation context, and accessible text relating to scripts
triggered by power state or time conditions are somewhat
inconsistent.

This change harmonizes the uses, based on the model in kcm_keys. The
preferred form is "command or script", which can be left out if already
very clear from context. The word "custom" is avoided, as the script may
not be custom from the user's point of view (e.g. OEM-supplied).
Other wording is also modeled after kcm_keys for consistency.