kcmodule: Accessible.Role on run condition button
addScriptCommandButton opens a menu with checkboxes that allow the user
to select for which conditions running a command or script should be
configurable. It is not exposed visually or through screen readers that
this button will open a menu.

This change explicitly sets the Accessible.Role to ButtonMenu, which
will indicate this to screen readers and automatically display a
downward pointing arrow on the button when using qqc2-desktop-style.