applets/batterymonitor: Make active while charging again
For Plasma 6, we changed this widget to become passive while plugged in
but charging. This can cause you to lose track of the charge level
during the time before the device is fully charged, and potentially make
bad decisions as a result of it (e.g. unplug your laptop and bring it
with you when it really needs more time to charge).

It hasn't been the most popular change of Plasma 6. It's not a bad idea
in general, but I feel like it's not the best behavior to apply to
everyone.

Let's bring back the old behavior. The widget already becomes passive
when fully charged, which should be space-saving enough.

BUG: 486781
FIXED-IN: 6.2.0