daemon: Use new OSD screenBrightnessChanged API
For the OSD service to present the brightness changes optimally
with potential popups on each relevant monitor, we try to match
ScreenBrightnessController displays to KScreen outputs.
If a matching KScreen output is found, some of its data is
included on top of the basic brightness change info.

Depends on the new KScreen::Edid::rawData() method from
https://invent.kde.org/plasma/libkscreen/-/merge_requests/196.