daemon: Mark read-only charge thresholds as unsupported (Linux)
Some systems will provide a start threshold value, but cannot set it
because the firmware derives it from the end threshold. In this case,
we're better off just marking it as unsupported altogether.

For a more in-depth explanation with Apple Macs as example, see:
https://gitlab.freedesktop.org/upower/upower/-/merge_requests/208#note_2531122

This commit includes a subtle change to the getthresholds method:
Instead of returning the lowest end threshold and the highest
start threshold, it will now return the single start threshold
that corresponds to the lowest end threshold. Whether or not that
start threshold is supported by the given battery.