daemon: Move the new ScreenBrightness D-Bus API out of actions
Being an action ties the interface to a particular path and
interface name. It also ties the lifetime of the D-Bus API to
the isSupported() method, which in turn complicates moving the
actual profile-based brightness change action away from the
deprecated BrightnessControl class.

Instead, let's elevate the new interface to the same level as
PolicyAgent, initializing it unconditionally in PowerDevilApp and
giving it a nice D-Bus bus name / path / interface at
"org.kde.ScreenBrightness".