daemon: Port profile-based brightness changes away from legacy API
The relevant code is moved from the deprecated BrightnessControl
action to the new ScreenBrightnessControl action, where its
isSupported() result is now independent of whether the (deprecated)
BrightnessControl D-Bus interface should be exposed or not.

The new per-display implementation of profile-based brightness
changes will preserve legacy API behavior and limit changes
to internal displays if any are available. See code comment for
a more detailed explanation.

BrightnessControl is now only responsible for its D-Bus API,
and can be removed whenever we decide it's not worth keeping.