daemon: Introduce proper dimming API for ScreenBrightnessController
This allows us to correctly restore brightness for displays that
were disconnected when they were dimmed, and are reconnected later.
It doesn't quite fix restoring dimmed displays across reboots yet,
this will be an extension requiring Wayland protocol updates.

`ScreenBrightnessController` now keeps track of more data for each
connected display, enough to recognize the display after it appears
again later (possibly with a different display ID) and restore some
runtime state.

This allows us to store and restore values for a display's
nominal brightness, minimum brightness, and dimming ratio.

The dimming ratio is multiplied with the nominal brightness and
capped by the minimum brightness to produce a display's desired
actual brightness. Several callers can set a dimming ratio,
the lowest one will be used. For now, the DimDisplay action is
the only user; however, other dimming methods are anticipated
in the future.

Importantly, by being able to restore these values separately
from just a single hardware brightness value, we can tell if the
display is still in the expected state from before the disconnect.
If it is, we can continue using the earlier nominal brightness but
update the dimming ratio to the current runtime setting.
Unrecognized or changed displays can remain untouched, they can be
assumed with a 1.0 dimming ratio from their current brightness level.

Note that this API is meant only for temporary dimming from the
baseline of the current nominal brightness. It isn't meant for
profile-dependent brightness changes or for ambient light sensors
driving gradual, automatic brightness adjustments. We can have
dedicated APIs later for these use cases.

BUG: 452492