applets/brightness: Check if KeyboardBrightness action is available
Besides reacting to changes in the `Solid.PowerManagement` DBus service,
we also need to track changes in the `KeyboardBrightnessControl` action
(which translates into a specific DBus path) before trying to call or
connect to that path and fail.

This should fix race conditions on plasma login, where it is possible
to get the service registered but the specific action still pending.

CCBUG: 486067