applets/brightness: Check if KeyboardBrightness action is available
Besides reacting to changes in the `Solid.PowerManagement` DBus service,
we also need to track changes in the `KeyboardBrightnessControl` action
(which translates into a specific DBus path) before trying to call or
connect to that path and fail.

This doesn't fix the race conditions on plasma login, but allows the
applet to react live to changes in the action list.

CCBUG: 486067