[daemon]: skip warnings if batteries aren't present
When hotpluggable batteries are unplugged, the user receives a
critical battery level warning. This doesn't make sense, since
this happens as direct consequence to the unplugging, not to
batteries running low or critical.

This likely isn't even an issue on most devices in the wild,
since batteries are usually hard to remove without opening the
case. For devices that do have hotpluggable batteries, add a
check if the battery is present before showing a warning
notification.

Signed-off-by: Sebastian Kŭgler <sebas@kde.org>