actions/dimdisplay: Unconditionally restore the dimming ratio
Commit b6b563f7 greatly simplified the `DimDisplay` action by
providing a dimming ratio property in `ScreenBrightnessController`.
This allowed us to get rid of manual brightness setters here.

What the commit was missing is to take out the condition for
`profileSettings.useProfileSpecificDisplayBrightness()` entirely.
There is no more clash with the `ScreenBrightnessControl` action,
so we don't need to limit out setter calls.

Instead, always resetting the dimming ratio to 1.0 is the correct
thing to do when switching to a profile without dimming.