applets/batterymonitor: Assume default power profile is used if power profile is unset
Before this patch, `isInDefaultPowerProfile` was set to `false` when
power-profiles-daemon is not installed. This causes the batterymonitor
system tray entry to remain stuck in the active state, which is
rather annoying.

This happened because `defaultPowerProfile`, which is always set to a
non-empty string, and `activeProfile`, which is blank when the p-p-d
is missing, are not equal.

This patch sets `isInDefaultPowerProfile` to `true` when `activeProfile`
is unset. On a system without p-p-d installed, this fixes the behaviour
of the batterymonitor widget and correctly puts it into passive state.
On a system with p-p-d installed, this does not alter behaviour,
because it only triggers when `activeProfile` is blank.

BUG: 494506