kcm: More accurate inline description of "Suspend, then hibernate"
On desktop systems where no battery is present, the previous
description was indeed accurate. But on systems with battery,
systemd tries hard to estimate the remaining battery time and
switches to hibernation only when the charge is almost depleted.
By default, at least.

So for systems with power supply battery, improve the description.