applets/battery: Fix navigation keyboard cycle for the Up key
When on the first item of the battery list, trying to force focus
to the header item fails, probably because it is wrapped within a
Column.

Just setting it to `null` works in this case, since it automatically
completes the `KeyNavigation.down` chain, but in reverse. This is
the same trick already being used in the footer.

BUG: 489099
FIXED-IN: 6.3.1